refactor: to feature-sliced methodology
