+++ src/client.ts
@@ -1,0 +1,1 @@
+import './app/client';
