+++ src/entities/card/organisms/card-preview.tsx
