+++ src/entities/card/types.ts
