+++ src/features/search-bar/organisms/index.ts
@@ -1,0 +1,1 @@
+export { Searchbar } from './search-bar';
