--- src/features/searchbar/organisms/index.ts
@@ -1,1 +1,0 @@
-export { Searchbar } from './searchbar';
